+++ source/js/routes/app/client.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Router, Route } from 'react-router-dom';
-import AppContainer from '../../containers/root/appContainer';
+import AppContainer from './container';
 import history from '../../config/browserHistory';
 
 export default () => (
